💄 home page update
