+++ src/app/[locale]/page.tsx
@@ -1,94 +1,29 @@
 import type { Metadata } from 'next';
-import { draftMode } from 'next/headers';
-import Link from 'next/link';
-import { notFound } from 'next/navigation';
 
-import { ArticleHero, ArticleTileGrid } from '@src/components/features/article';
-import { Container } from '@src/components/shared/container';
-import TranslationsProvider from '@src/components/shared/i18n/TranslationProvider';
-import initTranslations from '@src/i18n';
-import { defaultLocale, locales } from '@src/i18n/config';
-import { PageBlogPostOrder } from '@src/lib/__generated/sdk';
-import { client, previewClient } from '@src/lib/client';
-
-interface LandingPageProps {
+interface HomePageProps {
   params: {
     locale: string;
   };
 }
 
-export async function generateMetadata({ params }: LandingPageProps): Promise<Metadata> {
-  const { isEnabled: preview } = draftMode();
-  const gqlClient = preview ? previewClient : client;
-  const landingPageData = await gqlClient.pageLanding({ locale: params.locale, preview });
-  const page = landingPageData.pageLandingCollection?.items[0];
-
-  const languages = Object.fromEntries(
-    locales.map(locale => [locale, locale === defaultLocale ? '/' : `/${locale}`]),
-  );
-  const metadata: Metadata = {
+export async function generateMetadata({}: HomePageProps): Promise<Metadata> {
+  return {
+    title: 'Homepage',
+    description: 'Welcome to our new text-based homepage.',
     alternates: {
       canonical: '/',
-      languages: languages,
     },
   };
-  if (page?.seoFields) {
-    metadata.title = page.seoFields.pageTitle;
-    metadata.description = page.seoFields.pageDescription;
-    metadata.robots = {
-      follow: !page.seoFields.nofollow,
-      index: !page.seoFields.noindex,
-    };
-  }
-
-  return metadata;
 }
 
-export default async function Page({ params: { locale } }: LandingPageProps) {
-  const { isEnabled: preview } = draftMode();
-  const { t, resources } = await initTranslations({ locale });
-  const gqlClient = preview ? previewClient : client;
-
-  const landingPageData = await gqlClient.pageLanding({ locale, preview });
-  const page = landingPageData.pageLandingCollection?.items[0];
-
-  if (!page) {
-    notFound();
-  }
-
-  const blogPostsData = await gqlClient.pageBlogPostCollection({
-    limit: 6,
-    locale,
-    order: PageBlogPostOrder.PublishedDateDesc,
-    where: {
-      slug_not: page?.featuredBlogPost?.slug,
-    },
-    preview,
-  });
-  const posts = blogPostsData.pageBlogPostCollection?.items;
-
-  if (!page?.featuredBlogPost || !posts) {
-    return;
-  }
-
+export default function Page({}: HomePageProps) {
   return (
-    <TranslationsProvider locale={locale} resources={resources}>
-      <Container>
-        <Link href={`/${page.featuredBlogPost.slug}`}>
-          <ArticleHero article={page.featuredBlogPost} />
-        </Link>
-      </Container>
-
-      {/* Tutorial: contentful-and-the-starter-template.md */}
-      {/* Uncomment the line below to make the Greeting field available to render */}
-      {/*<Container>*/}
-      {/*  <div className="my-5 bg-colorTextLightest p-5 text-colorBlueLightest">{page.greeting}</div>*/}
-      {/*</Container>*/}
-
-      <Container className="my-8  md:mb-10 lg:mb-16">
-        <h2 className="mb-4 md:mb-6">{t('landingPage.latestArticles')}</h2>
-        <ArticleTileGrid className="md:grid-cols-2 lg:grid-cols-3" articles={posts} />
-      </Container>
-    </TranslationsProvider>
+    <main className="p-8">
+      <h1 className="text-4xl font-bold">Homepage</h1>
+      <p className="mt-4 text-lg">
+        Welcome to our new localized homepage! This page is now a text-based landing page with no
+        blog content.
+      </p>
+    </main>
   );
 }
